requirements.txt
Flask==3.0.3
werkzeug==3.0.3
mysql-connector-python==9.5.0
Jinja2>=3.1
fastapi==0.115.5
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
requests==2.32.3
python-docx==1.1.2
openpyxl==3.1.5

